:package: Exports DequeueVec
