+++ size.json
@@ -1,10 +1,10 @@
 {
   "minified": {
-    "pretty": "6.45 kB",
-    "raw": 6449
+    "pretty": "7.57 kB",
+    "raw": 7567
   },
   "brotli": {
-    "pretty": "1.94 kB",
-    "raw": 1944
+    "pretty": "2.26 kB",
+    "raw": 2263
   }
 }
